+++ frontreact/CHANGELOG.md
@@ -4,9 +4,9 @@ The following tools were used to generate this project:
 - create-vite
 
 The following steps were used to generate this project:
-- Create react project with create-vite: `npm init --yes vite@latest frontend -- --template=react`.
+- Create react project with create-vite: `npm init --yes vite@latest frontreact -- --template=react`.
 - Updating `vite.config.js` with port.
-- Create project file (`frontend.esproj`).
+- Create project file (`frontreact.esproj`).
 - Create `launch.json` to enable debugging.
 - Add project to solution.
 - Write this file.